ISSUE home pu /send ip

+++ script/install_5d.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 clear
-echo MeshDash Install-Script V 1.00.40
+echo MeshDash Install-Script V 1.00.42
 echo
 echo Installation Von MeshDash SQL
 echo 
@@ -24,6 +24,73 @@ else
 fi
 echo "OK, es geht weiter mit der Installation..."
 sleep 2
+
+##################################################
+# Funktion, um das Verzeichnis /home/pi zu überprüfen
+check_home_pi() {
+    if [ "$(pwd)" == "/home/pi" ]; then
+        # Wenn wir bereits im Verzeichnis /home/pi sind
+        echo "Wir sind bereits im Verzeichnis /home/pi."
+        return 0
+    elif [ -d "/home/pi" ]; then
+        # Wenn /home/pi existiert, aber nicht das aktuelle Verzeichnis ist
+        echo "Das Verzeichnis /home/pi existiert und wir haben Schreibrechte."
+        return 1
+    else
+        # Wenn /home/pi nicht existiert
+        echo "Das Verzeichnis /home/pi existiert nicht. Wir werden es nun erstellen."
+        sudo mkdir -p /home/pi
+        if [ $? -eq 0 ]; then
+            echo "Verzeichnis /home/pi erfolgreich erstellt."
+            return 1
+        else
+            echo "Fehler beim Erstellen des Verzeichnisses /home/pi."
+            exit 1
+        fi
+    fi
+}
+
+# Dateien, die kopiert werden sollen
+FILE1="install_5d.sh"
+FILE2="update_5d.sh"
+ZIPFILE=$(ls *.zip)
+
+# Funktion zum Kopieren der Dateien
+copy_files() {
+
+    echo "Kopiere Dateien nach /home/pi..."
+
+    sudo cp "$FILE1" /home/pi/
+    sudo cp "$FILE2" /home/pi/
+
+    # Überprüfen, ob eine .zip Datei existiert
+    if [ -z "$ZIPFILE" ]; then
+        echo "Keine .zip Datei gefunden."
+        exit 1
+    else
+        sudo cp "$ZIPFILE" /home/pi/
+    fi
+
+    if [ $? -eq 0 ]; then
+        echo "Dateien erfolgreich kopiert."
+    else
+        echo "Fehler beim Kopieren der Dateien."
+        exit 1
+    fi
+}
+
+# Hauptlogik
+check_home_pi
+if [ $? -eq 0 ]; then
+    echo "Installationsvorgang kann fortgesetzt werden."
+      echo "Wechseln ins Verzeichnis /home/pi..."
+        cd /home/pi
+else
+    copy_files
+    echo "Haupt-Installation wird nun ausgeführt."
+      echo "Wechseln ins Verzeichnis /home/pi..."
+         cd /home/pi
+fi
 #############Apt Install php, php-sqlite3, lighttpd
 echo
 echo Installiere jetzt weitere notwendige Software
@@ -95,8 +162,10 @@ echo
 #######################################
 hostIp=$(hostname -I | awk '{print $1}')
 echo
-echo Stoppe checkmh Service da neue Version ggf. kopiert wird
-sudo systemctl stop checkmh.service
+if systemctl is-active --quiet checkmh.service; then
+  echo Stoppe checkmh Service da neue Version ggf. kopiert wird
+  sudo systemctl stop checkmh.service
+fi
 echo
 echo Lösche meshdash Verzeichnis und erzeuge es neu
 sudo rm -rf meshdash
@@ -148,4 +217,4 @@ sudo systemctl start checkmh.service
 echo
 echo FERTIG!
 echo
-echo "Starte nun Deinen Webbrowser und gib http://$hostIp ein."
+echo "Starte nun Deinen Webbrowser und gib http://$hostIp/5d ein."
